Simplify error handling
