+++ api/src/actions/debug.js
@@ -4,7 +4,8 @@ module.exports = {
   authenticate: true,
   execute: async (context, user, input) => {
     const { now, env, db } = context;
+    const message = 'Debug info returned successfully.';
 
-    return success({ now, env, user, input });
+    return success({ now, env, user, input }, message);
   }
 };
